requirements.txt
python-dotenv
openai
llama-index
streamlit
nltk
youtube_transcript_api
llama-hub

